fix: 운영 환경 테스트 API 노출 차단 및 조직 조회 인증 적용
